refactor(sct_events): split module into package
