+++ longevity_lwt_test.py
@@ -1,12 +1,27 @@
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation; either version 3 of the License, or
+# (at your option) any later version.
 #
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+#
+# See LICENSE for more details.
+#
+# Copyright (c) 2020 ScyllaDB
+
 # This is stress longevity test that runs light weight transactions in parallel with different node operations:
 # disruptive and not disruptive
 #
 # After the test is finished will be performed the data validation.
+
 import time
 
 from longevity_test import LongevityTest
-from sdcm.sct_events import EventsSeverityChangerFilter, DatabaseLogEvent, Severity
+from sdcm.sct_events.base import Severity
+from sdcm.sct_events.filters import EventsSeverityChangerFilter
+from sdcm.sct_events.database import DatabaseLogEvent
 from sdcm.utils.data_validator import LongevityDataValidator
