+++ performance_regression_alternator_test.py
@@ -14,7 +14,7 @@
 import contextlib
 
 from performance_regression_test import PerformanceRegressionTest
-from sdcm.group_common_events import ignore_operation_errors, ignore_alternator_client_errors
+from sdcm.sct_events.group_common_events import ignore_operation_errors, ignore_alternator_client_errors
 from sdcm.utils import alternator
