+++ performance_regression_lwt_test.py
@@ -1,9 +1,24 @@
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation; either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+#
+# See LICENSE for more details.
+#
+# Copyright (c) 2020 ScyllaDB
+
 import pprint
 
 from invoke.exceptions import UnexpectedExit, Failure
+
 from performance_regression_test import PerformanceRegressionTest
-from sdcm.group_common_events import ignore_operation_errors
 from sdcm.utils.decorators import log_run_info, retrying
+from sdcm.sct_events.group_common_events import ignore_operation_errors
+
 
 PP = pprint.PrettyPrinter(indent=2)
