+++ sdcm/ndbench_thread.py
@@ -1,3 +1,16 @@
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation; either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+#
+# See LICENSE for more details.
+#
+# Copyright (c) 2020 ScyllaDB
+
 import os
 import re
 import logging
@@ -6,11 +19,13 @@ import uuid
 from distutils.util import strtobool  # pylint: disable=import-error,no-name-in-module
 
 from sdcm.prometheus import nemesis_metrics_obj
-from sdcm.sct_events import NdbenchStressEvent, Severity
+from sdcm.sct_events.base import Severity
+from sdcm.sct_events.loaders import NdbenchStressEvent
 from sdcm.utils.common import FileFollowerThread
 from sdcm.remote import FailuresWatcher
 from sdcm.utils.docker_remote import RemoteDocker
 from sdcm.stress_thread import format_stress_cmd_error, DockerBasedStressThread
+
 
 LOGGER = logging.getLogger(__name__)
