+++ sdcm/sct_events/group_common_events.py
@@ -1,7 +1,23 @@
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation; either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+#
+# See LICENSE for more details.
+#
+# Copyright (c) 2020 ScyllaDB
+
 from contextlib import contextmanager, ExitStack
 
-from sdcm.sct_events import DbEventsFilter, Severity, DatabaseLogEvent, EventsSeverityChangerFilter, YcsbStressEvent, \
-    PrometheusAlertManagerEvent, EventsFilter
+from sdcm.sct_events.base import Severity
+from sdcm.sct_events.filters import DbEventsFilter, EventsSeverityChangerFilter, EventsFilter
+from sdcm.sct_events.loaders import YcsbStressEvent
+from sdcm.sct_events.database import DatabaseLogEvent
+from sdcm.sct_events.monitors import PrometheusAlertManagerEvent
 
 
 @contextmanager
