+++ sdcm/ycsb_thread.py
@@ -1,3 +1,16 @@
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation; either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+#
+# See LICENSE for more details.
+#
+# Copyright (c) 2020 ScyllaDB
+
 import os
 import random
 import re
@@ -8,7 +21,8 @@ import tempfile
 from textwrap import dedent
 
 from sdcm.prometheus import nemesis_metrics_obj
-from sdcm.sct_events import YcsbStressEvent, Severity
+from sdcm.sct_events.base import Severity
+from sdcm.sct_events.loaders import YcsbStressEvent
 from sdcm.remote import FailuresWatcher
 from sdcm.utils import alternator
 from sdcm.utils.common import FileFollowerThread
