+++ unit_tests/conftest.py
@@ -1,17 +1,31 @@
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation; either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+#
+# See LICENSE for more details.
+#
+# Copyright (c) 2020 ScyllaDB
+
 import os
 import logging
 import tempfile
-from pathlib import Path
 import collections
+from pathlib import Path
 
 import pytest
 
-from sdcm.prometheus import start_metrics_server
-from sdcm.utils.docker_remote import RemoteDocker
 from sdcm import wait
-from sdcm.utils.decorators import timeout
-from sdcm.sct_events import (start_events_device, stop_events_device)
 from sdcm.cluster import BaseNode
+from sdcm.prometheus import start_metrics_server
+from sdcm.utils.decorators import timeout
+from sdcm.utils.docker_remote import RemoteDocker
+from sdcm.sct_events.events_device import start_events_device, stop_events_device
+
 from unit_tests.dummy_remote import LocalNode, LocalScyllaClusterDummy
