+++ unit_tests/test_decode_backtrace.py
@@ -1,6 +1,16 @@
-from __future__ import print_function
+# This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation; either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+#
+# See LICENSE for more details.
+#
+# Copyright (c) 2020 ScyllaDB
 
-from __future__ import absolute_import
 import unittest
 import tempfile
 import logging
@@ -9,10 +19,9 @@ import os
 import json
 import queue
 
-
 from sdcm.cluster import Setup
-from sdcm.sct_events import start_events_device, stop_events_device
-from sdcm.sct_events import EVENTS_PROCESSES
+from sdcm.sct_events.events_device import start_events_device, stop_events_device
+from sdcm.sct_events.events_processes import EVENTS_PROCESSES
 
 from unit_tests.dummy_remote import DummyRemote
 from unit_tests.test_cluster import DummyNode
